+++ packages/itwin/map-layers-mcp/.gitignore
@@ -1,0 +1,2 @@
+dist/
+node_modules/
